clean and extract
