+++ src/main/python/plotlyst/view/widget/scene/agency.py
@@ -22,8 +22,8 @@ from typing import Dict, Optional, Tuple, Union
 
 import qtanim
 from PyQt6.QtCore import Qt, QEvent, pyqtSignal, QSize, QTimer, QMimeData
-from PyQt6.QtGui import QEnterEvent, QMouseEvent, QCursor, QDragEnterEvent, QDragLeaveEvent, QResizeEvent
-from PyQt6.QtWidgets import QWidget, QSlider, QGridLayout, QButtonGroup, QAbstractButton, QFrame
+from PyQt6.QtGui import QEnterEvent, QCursor, QDragEnterEvent, QDragLeaveEvent, QResizeEvent
+from PyQt6.QtWidgets import QWidget, QGridLayout, QButtonGroup, QAbstractButton, QFrame
 from overrides import overrides
 from qtanim import fade_in
 from qthandy import hbox, spacer, sp, bold, vbox, translucent, clear_layout, margins, vspacer, \
@@ -42,11 +42,10 @@ from plotlyst.events import NovelEmotionTrackingToggleEvent, \
 from plotlyst.service.cache import entities_registry
 from plotlyst.view.common import push_btn, label, fade_out_and_gc, tool_btn, action, shadow, frame, scroll_area, rows, \
     columns
-from plotlyst.view.generated.scene_goal_stakes_ui import Ui_GoalReferenceStakesEditor
 from plotlyst.view.icons import IconRegistry, avatars
 from plotlyst.view.layout import group
 from plotlyst.view.style.base import apply_white_menu, transparent_menu
-from plotlyst.view.widget.button import ChargeButton, SmallToggleButton, SelectorToggleButton, DotsMenuButton
+from plotlyst.view.widget.button import SelectorToggleButton, DotsMenuButton
 from plotlyst.view.widget.character.editor import EmotionEditorSlider
 from plotlyst.view.widget.characters import CharacterSelectorMenu
 from plotlyst.view.widget.confirm import confirmed
@@ -54,160 +53,7 @@ from plotlyst.view.widget.display import ArrowButton, SeparatorLineWithShadow, C
     MenuOverlayEventFilter, Icon, icon_text
 from plotlyst.view.widget.input import RemovalButton, TextEditBubbleWidget
 from plotlyst.view.widget.scene.conflict import ConflictIntensityEditor, CharacterConflictSelector
-
-
-class MotivationDisplay(QWidget, Ui_GoalReferenceStakesEditor):
-    def __init__(self, parent=None):
-        super().__init__(parent)
-        self.setupUi(self)
-        self._novel: Optional[Novel] = None
-        self._scene: Optional[Scene] = None
-        self._agenda: Optional[CharacterAgency] = None
-        bold(self.lblTitle)
-
-        self._sliders: Dict[Motivation, QSlider] = {
-            Motivation.PHYSIOLOGICAL: self.sliderPhysiological,
-            Motivation.SAFETY: self.sliderSecurity,
-            Motivation.BELONGING: self.sliderBelonging,
-            Motivation.ESTEEM: self.sliderEsteem,
-            Motivation.SELF_ACTUALIZATION: self.sliderActualization,
-            Motivation.SELF_TRANSCENDENCE: self.sliderTranscendence,
-        }
-
-        for slider in self._sliders.values():
-            slider.setEnabled(False)
-        translucent(self)
-
-    @overrides
-    def mouseReleaseEvent(self, event: QMouseEvent) -> None:
-        pass
-
-    def setNovel(self, novel: Novel):
-        self._novel = novel
-
-    def setScene(self, scene: Scene):
-        self._scene = scene
-
-    def setAgenda(self, agenda: CharacterAgency):
-        self._agenda = agenda
-        self._refresh()
-
-    def _refresh(self):
-        for slider in self._sliders.values():
-            slider.setValue(0)
-        for scene in self._novel.scenes:
-            if scene is self._scene:
-                break
-            for agenda in scene.agency:
-                if agenda.character_id and agenda.character_id == self._agenda.character_id:
-                    for mot, v in agenda.motivations.items():
-                        slider = self._sliders[Motivation(mot)]
-                        slider.setValue(slider.value() + v)
-
-
-class MotivationChargeLabel(QWidget):
-    def __init__(self, motivation: Motivation, simplified: bool = False, parent=None):
-        super().__init__(parent)
-        self._motivation = motivation
-        hbox(self, margin=0 if simplified else 1, spacing=0)
-        if simplified:
-            self._btn = tool_btn(IconRegistry.from_name(self._motivation.icon(), self._motivation.color()),
-                                 icon_resize=False, transparent_=True)
-        else:
-            self._btn = push_btn(IconRegistry.from_name(self._motivation.icon(), self._motivation.color()),
-                                 text=motivation.display_name(), icon_resize=False,
-                                 transparent_=True)
-        self._btn.setCursor(Qt.CursorShape.ArrowCursor)
-        decr_icon(self._btn, 2)
-
-        self._lblCharge = label('', description=True, italic=True, decr_font_diff=1)
-
-        self.layout().addWidget(self._btn)
-        self.layout().addWidget(self._lblCharge)
-
-    def setCharge(self, charge: int):
-        bold(self._btn, charge > 0)
-        if charge == 0:
-            self._lblCharge.clear()
-        else:
-            self._lblCharge.setText(f'+{charge}')
-
-
-class MotivationCharge(QWidget):
-    charged = pyqtSignal(int)
-    MAX_CHARGE: int = 5
-
-    def __init__(self, motivation: Motivation, parent=None):
-        super().__init__(parent)
-        hbox(self)
-        self._motivation = motivation
-        self._charge = 0
-
-        self._label = MotivationChargeLabel(self._motivation)
-        self._posCharge = ChargeButton(positive=True)
-        self._posCharge.clicked.connect(lambda: self._changeCharge(1))
-        self._negCharge = ChargeButton(positive=False)
-        self._negCharge.clicked.connect(lambda: self._changeCharge(-1))
-        self._negCharge.setHidden(True)
-
-        self.layout().addWidget(self._label)
-        self.layout().addWidget(spacer())
-        self.layout().addWidget(self._negCharge)
-        self.layout().addWidget(self._posCharge)
-
-    def setValue(self, value: int):
-        self._charge = min(value, self.MAX_CHARGE)
-        self._update()
-
-    def _changeCharge(self, charge: int):
-        self._charge += charge
-        self._update()
-
-        self.charged.emit(self._charge)
-
-    def _update(self):
-        self._label.setCharge(self._charge)
-        if self._charge == 0:
-            self._negCharge.setHidden(True)
-        else:
-            self._negCharge.setVisible(True)
-        if self._charge == self.MAX_CHARGE:
-            self._posCharge.setHidden(True)
-        else:
-            self._posCharge.setVisible(True)
-
-
-class MotivationEditor(QWidget):
-    motivationChanged = pyqtSignal(Motivation, int)
-
-    def __init__(self, parent=None):
-        super().__init__(parent)
-
-        vbox(self)
-        self.layout().addWidget(label("Does the character's motivation change?"))
-
-        self._editors: Dict[Motivation, MotivationCharge] = {}
-        self._addEditor(Motivation.PHYSIOLOGICAL)
-        self._addEditor(Motivation.SAFETY)
-        self._addEditor(Motivation.BELONGING)
-        self._addEditor(Motivation.ESTEEM)
-        self._addEditor(Motivation.SELF_ACTUALIZATION)
-        self._addEditor(Motivation.SELF_TRANSCENDENCE)
-
-    def _addEditor(self, motivation: Motivation):
-        wdg = MotivationCharge(motivation)
-        self._editors[motivation] = wdg
-        wdg.charged.connect(partial(self.motivationChanged.emit, motivation))
-        self.layout().addWidget(wdg)
-
-    def reset(self):
-        for editor in self._editors.values():
-            editor.setValue(0)
-
-    def setMotivations(self, motivations: Dict[Motivation, int]):
-        self.reset()
-        for mot, v in motivations.items():
-            self._editors[mot].setValue(v)
+from plotlyst.view.widget.scene.motivation import MotivationDisplay, MotivationEditor, MotivationChargeLabel
 
 
 class SceneAgendaEmotionEditor(QWidget):
@@ -262,7 +108,7 @@ class SceneAgendaEmotionEditor(QWidget):
 class SceneAgendaMotivationEditor(QWidget):
     motivationChanged = pyqtSignal(Motivation, int)
 
-    def __init__(self, parent=None):
+    def __init__(self, novel: Novel, scene: Scene, agency: CharacterAgency, parent=None):
         super().__init__(parent)
         hbox(self)
         sp(self).h_max()
@@ -287,40 +133,20 @@ class SceneAgendaMotivationEditor(QWidget):
         self.layout().addWidget(self._icon)
         self.layout().addWidget(self._wdgLabels)
 
-        self.reset()
-
-    def setNovel(self, novel: Novel):
         self._motivationDisplay.setNovel(novel)
-
-    def setScene(self, scene: Scene):
         self._motivationDisplay.setScene(scene)
+        self._motivationDisplay.setAgenda(agency)
 
-    def setAgenda(self, agenda: CharacterAgency):
-        self._motivationDisplay.setAgenda(agenda)
-
-        if agenda.motivations:
+        if agency.motivations:
             values: Dict[Motivation, int] = {}
-            for k, v in agenda.motivations.items():
+            for k, v in agency.motivations.items():
                 motivation = Motivation(k)
                 values[motivation] = v
 
-            self.setValues(values)
-        else:
-            self.reset()
+            self._motivationEditor.setMotivations(values)
 
-    def reset(self):
-        self._motivationEditor.reset()
-
-        self._labels.clear()
-        clear_layout(self._wdgLabels)
-
-    def setValues(self, motivations: Dict[Motivation, int]):
-        self._motivationEditor.setMotivations(motivations)
-
-        self._labels.clear()
-        clear_layout(self._wdgLabels)
-        for mot, v in motivations.items():
-            self._updateLabels(mot, v)
+            for mot, v in values.items():
+                self._updateLabels(mot, v)
 
     def _valueChanged(self, motivation: Motivation, value: int):
         self.motivationChanged.emit(motivation, value)
@@ -429,12 +255,6 @@ class SceneAgendaConflictEditor(QWidget):
         conflictSelector = CharacterConflictSelector(self._novel, self._scene, self._agenda)
         conflictSelector.conflictSelected.connect(self._conflictSelected)
         self._wdgConflicts.layout().addWidget(conflictSelector)
-
-
-class _CharacterStateToggle(SmallToggleButton):
-    def __init__(self, type_: StoryElementType, parent=None):
-        super().__init__(parent)
-        self.type = type_
 
 
 class _CharacterChangeSelectorToggle(SelectorToggleButton):
@@ -888,11 +708,8 @@ class CharacterMotivationChange(CharacterChangeBubble):
         super().__init__(element, parent)
         self._agency = agency
 
-        self.motivationEditor = SceneAgendaMotivationEditor()
+        self.motivationEditor = SceneAgendaMotivationEditor(novel, scene, agency)
         self.motivationEditor.motivationChanged.connect(self._motivationChanged)
-        self.motivationEditor.setNovel(novel)
-        self.motivationEditor.setScene(scene)
-        self.motivationEditor.setAgenda(agency)
         self.addBottomWidget(self.motivationEditor)
 
     def _motivationChanged(self, motivation: Motivation, value: int):
